[debug]user

+++ Code/DMP-service/dmp/api/file.py
@@ -5,9 +5,9 @@
 import os
 
 import pandas as pd
-from flask import Blueprint, jsonify, request, current_app
+from flask import Blueprint,  request, current_app
 from dmp.utils import resp_hanlder, uuid_str
-from dmp.models import FormDownload
+from dmp.models import Forms
 
 file = Blueprint("file", __name__)
 
@@ -69,8 +69,7 @@ def success(desc):
     return resp_hanlder(result={"filename": finally_filename})
 
 
-class FormDownload(object):
-    pass
+
 
 
 @file.route("/dlcomplete/", methods=["GET"], defaults={"desc": {"interface_name": "文件下载完成","is_permission": False,"permission_belong": None}})
@@ -78,9 +77,13 @@ def dlcomplete(desc):
     if request.method == "GET":
         try:
             form_id = request.json.get("form_id")
-            form_ = FormDownload.get(form_id)
-            rm_filepath = form_.filepath
-            os.remove(rm_filepath)
+            form_ = Forms.get(form_id)
+            rm_filepath = form_.info_form.filepath
+            if os.path.exists(rm_filepath):
+                os.remove(rm_filepath)
             return resp_hanlder(result="OK")
+            dlform = form_.info_form
+            dlform.ftp_url = "已失效！"
+            dlform.put()
         except Exception as err:
             return resp_hanlder(result="OK")
